fraud_transaction_detector-main/requirements.txt
asgiref==3.9.1
Django==5.1.4
django-widget-tweaks==1.5.0
dnspython==2.7.0
et_xmlfile==2.0.0
joblib==1.5.2
numpy==2.3.2
openpyxl==3.1.5
pandas==2.2.2
pillow==11.0.0
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2025.2
scikit-learn==1.7.1
scipy==1.16.1
six==1.17.0
sqlparse==0.5.2
threadpoolctl==3.6.0
tzdata==2025.2
whitenoise==6.8.2
xlrd==1.2.0
gunicorn==23.0.0
